pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biology-learning-agent"
version = "0.1.0"
description = "生物学习智能体 - 通过多轮迭代优化生物学试题答案"
readme = "README.md"
authors = [{name = "Biology Learning Team"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Topic :: Education :: Computer Aided Instruction (CAI)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.9"
dependencies = [
    "python-docx>=0.8.11",
    "requests>=2.31.0",
    "pydantic>=2.5.0",
    "pytest>=7.4.0",
    "jsonschema>=4.20.0",
    "pillow>=10.0.0",  # 图片处理支持
    "tenacity>=8.2.0", # 重试逻辑（可选）
    "pdf2image>=1.16.3",  # PDF转图片
    "pymupdf>=1.23.8",    # PDF文本提取
    "reportlab>=4.0.0",   # PDF生成（测试用）
]

[project.optional-dependencies]
dev = [
    "mypy>=1.8.0",
    "black>=24.0.0",
    "isort>=5.13.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.6.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/biology-learning-agent"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true